requirements.txt
Pillow==9.2.0
pytesseract==0.3.10
PyYAML==6.0
wxPython==4.1.0
xmltodict==0.12.0
